Change approach
The purpose of this exercise is to gain
some x86 ASM fluency and to document my
learning process. Initially I was planning
on printing out the box for the snake game
by individually piecing it together but
I came to the realisation that this is not
going to work. Next step is to use a 2D array.